refactor: factorize reduced-effects iridescent layers and drop legacy chrome
