+++ src/components/BarY2KOrnament.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import type { FC, CSSProperties } from 'react';
 import { LiquidMetal } from '@paper-design/shaders-react';
+import IridescentMaskedLayers from './IridescentMaskedLayers';
 
 
 // Default ornament size for bar overlay
@@ -65,18 +66,6 @@ const BarY2KOrnament: FC<Props> = ({
       ? { right: 0, transform: `translateX(50%) translateY(${yOffset}%)` }
       : { left: 0, transform: `translateX(-50%) translateY(${yOffset}%)` };
 
-  const maskBaseStyle: CSSProperties = {
-    position: 'absolute',
-    inset: 0,
-    WebkitMaskImage: `url(${src})`,
-    maskImage: `url(${src})`,
-    WebkitMaskRepeat: 'no-repeat',
-    maskRepeat: 'no-repeat',
-    WebkitMaskSize: '100% 100%',
-    maskSize: '100% 100%',
-    transformOrigin: 'center',
-  };
-
   const staticScaleStyle: CSSProperties = staticMode
     ? {
         transform: 'scale(0.93)',
@@ -98,57 +87,7 @@ const BarY2KOrnament: FC<Props> = ({
     >
       {staticMode ? (
         <div style={{ position: 'relative', width: '100%', height: '100%', ...staticScaleStyle }}>
-          <img
-            src={src}
-            alt=""
-            aria-hidden="true"
-            style={{
-              position: 'absolute',
-              inset: 0,
-              width: '100%',
-              height: '100%',
-              objectFit: 'fill',
-              transform: flip ? 'scaleX(-1)' : undefined,
-              filter: 'grayscale(0.35) brightness(1.08) contrast(1.28) saturate(1.2)',
-              opacity: 0.96,
-            }}
-          />
-          <div
-            aria-hidden="true"
-            style={{
-              ...maskBaseStyle,
-              transform: flip ? 'scaleX(-1)' : undefined,
-              background: 'linear-gradient(112deg, #ffffff 0%, #d6daea 18%, #8f86b0 34%, #72f3ff 50%, #ff6dd7 65%, #f9fbff 82%, #8afcff 100%)',
-              backgroundSize: '200% 200%',
-              mixBlendMode: 'screen',
-              opacity: 0.72,
-              animation: 'y2k-iridescent-shift 9s ease-in-out infinite',
-            }}
-          />
-          <div
-            aria-hidden="true"
-            style={{
-              ...maskBaseStyle,
-              transform: flip ? 'scaleX(-1)' : undefined,
-              background: 'conic-gradient(from 210deg at 50% 45%, #ffffff 0deg, #66f2ff 70deg, #a56bff 135deg, #ff58c5 220deg, #f7fcff 300deg, #ffffff 360deg)',
-              backgroundSize: '165% 165%',
-              mixBlendMode: 'color-dodge',
-              opacity: 0.52,
-              filter: 'blur(0.25px) saturate(1.25)',
-              animation: 'y2k-prism-drift 11s linear infinite',
-            }}
-          />
-          <div
-            aria-hidden="true"
-            style={{
-              ...maskBaseStyle,
-              transform: flip ? 'scaleX(-1)' : undefined,
-              background: 'linear-gradient(180deg, rgba(255,255,255,0.88) 0%, rgba(255,255,255,0) 30%, rgba(0,0,0,0.18) 100%)',
-              mixBlendMode: 'soft-light',
-              opacity: 0.55,
-              animation: 'y2k-specular-pulse 7s ease-in-out infinite',
-            }}
-          />
+          <IridescentMaskedLayers maskSrc={src} flip={flip} />
         </div>
       ) : (
         mask && (
